Toujours dans l'implementation des services et de la modification des erreurs ou des mauvais bout de code decouvert dans les parties anterieures. Presentement on est deja au niveau des services du provider notamment les comptes capsules et damages qui seront implemente apres ce push

+++ src/main/java/com/c2psi/businessmanagement/dtos/pos/pos/PosCapsuleOperationDto.java
@@ -9,6 +9,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -19,7 +20,7 @@ public class PosCapsuleOperationDto {
 
     @NotNull(message = "The number in mouvement cannot be null")
     @Positive(message = "The number in mouvement must be positive")
-    Integer poscsoNumberinmvt;
+    BigDecimal poscsoNumberinmvt;
     /******************************
      * Relation between entities  *
      * ****************************/
